+++ src/num/niche/mem/mod.rs
@@ -5,6 +5,7 @@
 
 #[cfg(test)]
 mod tests;
+// mod asm_validation;
 
 mod non_value;
 pub use non_value::*;